+++ src/sql/connection/connection.py
@@ -245,6 +245,7 @@ class ConnectionManager:
             conn = cls.connections.get(descriptor) or cls.connections.get(
                 descriptor.lower()
             )
+
         if not conn:
             raise exceptions.RuntimeError(
                 "Could not close connection because it was not found amongst these: %s"
@@ -257,7 +258,8 @@ class ConnectionManager:
             cls.connections.pop(
                 str(conn.metadata.bind.url) if IS_SQLALCHEMY_ONE else str(conn.url)
             )
-            conn.connection.close()
+
+        conn.close()
 
     @classmethod
     def connections_table(cls):
@@ -526,6 +528,13 @@ class SQLAlchemyConnection(AbstractConnection):
     def connection(self):
         """Returns the SQLAlchemy connection object"""
         return self._connection_sqlalchemy
+
+    def close(self):
+        super().close()
+
+        # NOTE: in SQLAlchemy 2.x, we need to call engine.dispose() to completely
+        # close the connection, calling connection.close() is not enough
+        self.connection.engine.dispose()
 
     @classmethod
     @modify_exceptions