+++ src/tests/conftest.py
@@ -4,13 +4,13 @@ import urllib.request
 from pathlib import Path
 
 import pytest
-from IPython.core.interactiveshell import InteractiveShell
 
 
 from sql.magic import SqlMagic, RenderMagic
 from sql.magic_plot import SqlPlotMagic
 from sql.magic_cmd import SqlCmdMagic
 from sql.connection import ConnectionManager
+from sql._testing import TestingShell
 from sql import connection
 
 PATH_TO_TESTS = Path(__file__).absolute().parent
@@ -49,6 +49,7 @@ def chinook_db():
     return str(path)
 
 
+# TODO: this is legacy code, we need to remove it
 def runsql(ip_session, statements):
     if isinstance(statements, str):
         statements = [statements]
@@ -64,21 +65,6 @@ def clean_conns():
     yield
 
 
-class TestingShell(InteractiveShell):
-    """
-    A custom InteractiveShell that raises exceptions instead of silently suppressing
-    them.
-    """
-
-    def run_cell(self, *args, **kwargs):
-        result = super().run_cell(*args, **kwargs)
-
-        if result.error_in_exec is not None:
-            raise result.error_in_exec
-
-        return result
-
-
 @pytest.fixture
 def ip_empty():
     c = Config()
@@ -86,26 +72,6 @@ def ip_empty():
     # which leads to "too many open files" error when running tests; this setting
     # disables the history recording.
     # https://ipython.readthedocs.io/en/stable/config/options/terminal.html#configtrait-HistoryAccessor.enabled
-    c.HistoryAccessor.enabled = False
-    ip_session = InteractiveShell(config=c)
-
-    ip_session.register_magics(SqlMagic)
-    ip_session.register_magics(RenderMagic)
-    ip_session.register_magics(SqlPlotMagic)
-    ip_session.register_magics(SqlCmdMagic)
-
-    # there is some weird bug in ipython that causes this function to hang the pytest
-    # process when all tests have been executed (an internal call to gc.collect()
-    # hangs). This is a workaround.
-    ip_session.displayhook.flush = lambda: None
-
-    yield ip_session
-    ConnectionManager.close_all()
-
-
-@pytest.fixture
-def ip_empty_testing():
-    c = Config()
     c.HistoryAccessor.enabled = False
     ip_session = TestingShell(config=c)